TH-1225 : 레거시 가게 상세 전용 죽은 코드 제거
- v1 가게 상세 스크린 API(fetchStoreScreen)·사장님 가게 상세 API(fetchBossStoreDetail) 제거
- 레거시 상세 화면만 사용하던 Presentation 모델 14종 제거 (StoreDetailData/BossStoreDetailData 클러스터)
- 관련 응답·입력 모델 제거 (StoreScreenResponse, BossStoreDetailResponse, BossStoreResponse, BossStoreMenuResponse, StoreAccountNumberResponse, FetchBossStoreDetailInput)
- 공유 타입 ActivitiesStatus는 별도 파일로 분리 유지
- ReviewListViewModel의 낡은 주석 정리
