--- Modules/Core/Model/Sources/Domain/Input/FetchBossStoreDetailInput.swift
@@ -1,17 +1,0 @@
-import Foundation
-
-public struct FetchBossStoreDetailInput: Encodable {
-    public let storeId: String
-    public let latitude: Double
-    public let longitude: Double
-
-    public init(
-        storeId: String,
-        latitude: Double,
-        longitude: Double
-    ) {
-        self.storeId = storeId
-        self.latitude = latitude
-        self.longitude = longitude
-    }
-}
