--- Modules/Core/Model/Sources/Domain/Response/BossStoreDetailResponse.swift
@@ -1,14 +1,0 @@
-import Foundation
-
-public struct BossStoreDetailResponse: Decodable {
-    public let store: BossStoreResponse
-    public let distanceM: Int
-    public let openStatus: StoreOpenResponse
-    public let favorite: StoreFavoriteResponse
-    public let feedbacks: [FeedbackCountWithRatioResponse]
-    // 서버 수정 필요
-    public let newsPosts: ContentsWithCursorWithTotalCountResponse<PostResponse>
-    public let tags: StoreTagResponse
-    public let coupons: ContentListStoreCouponSimpleResponse
-    public let reviews: ContentsWithCursorWithTotalCountResponse<StoreReviewWithWriterResponse>
-}
