--- Modules/Core/Model/Sources/Domain/Response/BossStoreMenuResponse.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-public struct BossStoreMenuResponse: Decodable {
-    public let name: String
-    public let price: Int
-    public let imageUrl: String?
-}
