--- Modules/Core/Model/Sources/Domain/Response/StoreAccountNumberResponse.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-public struct StoreAccountNumberResponse: Decodable {
-    public let bank: BankResponse
-    public let accountHolder: String
-    public let accountNumber: String
-    public let description: String?
-}
