--- Modules/Core/Model/Sources/Presentation/BossStoreDetailData.swift
@@ -1,62 +1,0 @@
-import Foundation
-
-public struct BossStoreDetailData {
-    public var overview: StoreDetailOverview
-    public let info: BossStoreInfo
-    public let coupons: [StoreCouponSimpleResponse]
-    public let menus: [BossStoreMenu]
-    public var workdays: [BossStoreAppearanceDay]
-    public var feedbacks: [FeedbackCountWithRatioResponse]
-    public var store: BossStoreResponse
-    public let post: PostResponse?
-    public let totalPostCount: Int?
-    public let totalReviewCount: Int
-    public var reviews: [StoreDetailReview]
-    public let isVerifiedStore: Bool
-
-    public init(response: BossStoreDetailResponse) {
-        self.overview = StoreDetailOverview(
-            categories: response.store.categories,
-            repoterName: "",
-            storeName: response.store.name,
-            isNew: response.tags.isNew,
-            totalVisitSuccessCount: 0,
-            reviewCount: response.reviews.cursor.totalCount,
-            distance: response.distanceM,
-            location: response.store.location,
-            address: response.store.address.fullAddress,
-            isFavorited: response.favorite.isFavorite,
-            subscribersCount: response.favorite.totalSubscribersCount,
-            isBossStore: true,
-            snsUrl: response.store.snsUrl,
-            introduction: response.store.introduction,
-            hasIssuableCoupon: response.tags.hasIssuableCoupon,
-            uniqueContributorCount: nil
-        )
-        self.info = BossStoreInfo(
-            updatedAt: response.store.updatedAt,
-            snsUrl: response.store.snsUrl,
-            introduction: response.store.introduction,
-            images: response.store.representativeImages,
-            accountInfos: response.store.accountNumbers.map { StoreAccountNumber(response: $0 )},
-            contactsNumbers: response.store.contactsNumbers
-        )
-        self.menus = response.store.menus.map { BossStoreMenu(response: $0) }
-        self.coupons = response.coupons.contents.map { $0 }
-        self.workdays = response.store.appearanceDays.map {
-            BossStoreAppearanceDay(response: $0)
-        }
-        self.feedbacks = response.feedbacks
-        self.store = response.store
-        if let post = response.newsPosts.contents.first {
-            self.post = post
-            self.totalPostCount = response.newsPosts.cursor.totalCount
-        } else {
-            self.post = nil
-            self.totalPostCount = nil
-        }
-        self.totalReviewCount = response.reviews.cursor.totalCount
-        self.reviews = response.reviews.contents.map { StoreDetailReview(response: $0) }
-        self.isVerifiedStore = response.tags.isVerifiedStore ?? false
-    }
-}
